refactor: Car의 원시값 name을 CarName으로 포장
